+++ web/static/react/assets/BackupPage-BJkK06y4.js
@@ -1,0 +1,1 @@
+import{t as e}from"./vendor-react-2_-HNKNn.js";import{t}from"./ImpostazioniPage-d4sGlFqX.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as BackupPage};